indexer/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nweb-indexer"
version = "0.1.0"
description = "nweb indexer - Watches blockchain for scan submissions and indexes IPFS bundles"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "web3>=6.15.1",
    "eth-abi>=4.2.1",
    "eth-utils>=2.3.1",
    "ipfshttpclient>=0.8.0a2",
    "aiohttp>=3.9.1",
    "httpx>=0.25.2",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "alembic>=1.13.1",
    "pydantic>=2.5.2",
    "msgpack>=1.0.7",
    "ujson>=5.9.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "asyncio-mqtt>=0.16.1",
    "aiofiles>=23.2.1",
]

[project.scripts]
nweb-indexer = "indexer.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "S101", "S104", "S105", "S106", "S107", "B008", "PLR0913", "PLR0912"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["web3.*", "eth_utils.*", "ipfshttpclient.*", "aiohttp.*"]
ignore_missing_imports = true
